Moved uip cut generation to Utils.
This required hoisting all types (as is appropriate) into Funsat.Types,
and adjusting imports.  The trade-off is that Funsat.Types now exports
some internal types which were hidden before.  Get over it.

Also, some of the helper functions that were in Funsat.Types (who knows why),
are in Funsat.Utils now.

This commit also removes all warnings (of those chosen, which certainly
is not -Wall).